Fixed several comparison of signed and unsigned integers.

+++ src/ccnl-core.c
@@ -255,7 +255,7 @@ ccnl_interface_enqueue(void (tx_done)(void*, int, int), struct ccnl_face_s *f,
 
     DEBUGMSG_CORE(TRACE, "enqueue interface=%p buf=%p len=%d (qlen=%d)\n",
                   (void*)ifc, (void*)buf,
-                  buf ? buf->datalen : -1, ifc ? ifc->qlen : -1);
+                  buf ? (int) buf->datalen : -1, ifc ? ifc->qlen : -1);
 
     if (ifc->qlen >= CCNL_MAX_IF_QLEN) {
         DEBUGMSG_CORE(WARNING, "  DROPPING buf=%p\n", (void*)buf);
